+++ packages/stylelint-plugin/src/linters/lintFunctions.ts
@@ -1,8 +1,8 @@
-import { Declaration, Root } from "postcss"
+import { Declaration } from "postcss"
 import postcssValueParser from "postcss-value-parser"
 import { invalidRgbaUsage, unmigratableDeclarationMessage } from "../errors"
 import { transformDecl } from "../functionTransformer"
-import { KaizenToken, kaizenTokensByName } from "../kaizenTokens"
+import { kaizenTokensByName } from "../kaizenTokens"
 import { containsUnmigratableFunction } from "../patterns"
 import { Options } from "../types"
 import { variablePrefixForLanguage } from "../utils"
