+++ packages/stylelint-plugin/src/linters/lintImports.ts
@@ -1,8 +1,7 @@
 import { atRule, Root } from "postcss"
 import { missingRequiredKaizeImport, unnecessaryKaizenImport } from "../errors"
-import { KaizenToken } from "../kaizenTokens"
 import { quotesPattern } from "../patterns"
-import { Options } from "../types"
+import { Options, KaizenToken } from "../types"
 import { walkKaizenTokens } from "../walkers"
 
 const getCurrentImports = (stylesheet: Root) => {
